refactor(stack-component): pass props from page component
