+++ src/app/global.scss
@@ -536,7 +536,7 @@ pre {
 
 /* Start Skills */
 .skills {
-  @apply mx-auto max-w-[1000px];
+  @apply mx-auto max-w-[1000px] py-14 md:py-0;
 
   min-height: 100dvh;
 
@@ -554,4 +554,61 @@ pre {
       content: '';
     }
   }
+
+  &_grid {
+    @apply grid grid-cols-1 gap-4 sm:grid-cols-2 lg:grid-cols-3;
+  }
+
+  &_item {
+    @apply rounded-lg border border-secondary-dark bg-primary-dark p-4;
+  }
+
+  &_header {
+    @apply mb-2 flex items-center;
+  }
+
+  &_icon {
+    @apply mr-2;
+  }
+
+  &_name {
+    @apply text-lg font-bold text-lightest-slate;
+  }
+
+  &_content {
+    @apply relative pt-1;
+  }
+
+  &_frequency {
+    @apply mb-2 flex items-center justify-between;
+  }
+
+  &_frequency-text {
+    @apply inline-block rounded-full border border-secondary-dark bg-primary-dark px-2 py-1 text-xs font-semibold uppercase text-secondary-dark;
+  }
+
+  &_frequency-percentage {
+    @apply inline-block text-xs font-semibold text-secondary-dark;
+  }
+
+  &_bar {
+    @apply mb-4 flex h-2 overflow-hidden rounded border border-secondary-dark bg-primary-dark text-xs;
+  }
+
+  &_bar-fill {
+    @apply flex flex-col justify-center whitespace-nowrap bg-secondary-dark text-center text-white shadow-none;
+  }
+
+  &_other-tech {
+    @apply mb-4 mt-8 text-center text-xl font-bold text-lightest-slate sm:text-2xl;
+  }
+
+  &_other-tech-container {
+    @apply flex flex-wrap justify-center gap-2;
+  }
+
+  &_other-tech-item {
+    @apply rounded-full bg-secondary-dark px-3 py-1 text-sm;
+  }
 }
+/* End Skills */
